+++ src/client/modules/Investments/Projects/Details/EditRecipientCompany/EditRecipientCompany.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { useParams } from 'react-router-dom'
 
-import { DefaultLayout } from '../../../components'
-import Task from '../../../components/Task'
+import { DefaultLayout } from '../../../../../components'
+import Task from '../../../../../components/Task'
 import { RECIPIENT_COMPANY_ID, TASK_UPDATE_RECIPIENT_COMPANY } from './state'
-import urls from '../../../../lib/urls'
-import { InvestmentResource } from '../../../components/Resource'
+import urls from '../../../../../../lib/urls'
+import { InvestmentResource } from '../../../../../components/Resource'
 import { checkIfRecipientCompanyExists } from './transformers'
-import { buildProjectBreadcrumbs } from '../utils'
+import { buildProjectBreadcrumbs } from '../../../utils'
 
 const checkIfUpdatingOrRemoving = (checkCompanyId, ukCompany) =>
   checkCompanyId
