+++ src/client/modules/Investments/Projects/Details/state.js
@@ -1,0 +1,10 @@
+export const TASK_EDIT_INVESTMENT_PROJECT_REQUIREMENTS =
+  'TASK_EDIT_INVESTMENT_PROJECT_REQUIREMENTS'
+export const TASK_EDIT_INVESTMENT_PROJECT_SUMMARY =
+  'TASK_EDIT_INVESTMENT_PROJECT_SUMMARY'
+export const TASK_EDIT_INVESTMENT_PROJECT_VALUE =
+  'TASK_EDIT_INVESTMENT_PROJECT_VALUE'
+
+export const state2props = (state) => ({
+  currentAdviserId: state.currentAdviserId,
+})
