+++ src/client/modules/Investments/Projects/Team/state.js
@@ -1,0 +1,5 @@
+export const TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER =
+  'TASK_SAVE_CLIENT_RELATIONSHIP_MANAGER'
+export const TASK_EDIT_PROJECT_TEAM_MEMBERS = 'TASK_EDIT_PROJECT_TEAM_MEMBERS'
+export const TASK_SAVE_INVESTMENT_PROJECT_MANAGERS =
+  'TASK_SAVE_INVESTMENT_PROJECT_MANAGERS'
