+++ src/client/modules/Investments/Projects/transformers.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import { Button, Link } from 'govuk-react'
-import { get } from 'lodash'
 
-import { transformDateObjectToDateString } from '../../../transformers'
 import { OPTION_NO, OPTION_YES } from '../../../../apps/constants'
 import urls from '../../../../lib/urls'
 import {
@@ -18,13 +16,8 @@ import {
   STAGE_VERIFY_WIN,
   STAGE_WON,
 } from './constants'
-import { transformArray } from '../../Companies/CompanyInvestments/LargeCapitalProfile/transformers'
 import { format } from '../../../utils/date'
 import { BLACK, GREY_3 } from '../../../utils/colours'
-import labels from '../../Companies/CollectionList/labels'
-import { addressToString } from '../../../utils/addresses'
-import { formatMediumDateTime } from '../../../utils/date'
-import { transformArrayToObject } from '../../../components/Form/elements/FieldAddAnother/utils'
 
 export const checkIfItemHasValue = (item) => (item ? item : null)
 
@@ -42,263 +35,14 @@ export const transformObjectForTypeahead = (object) =>
       }
     : null
 
-export const transformAdviserForAPI = (adviser) =>
-  adviser ? adviser.value : null
-
-export const transformObjectForTable = (object, emptyText = '') =>
-  object ? object.name : emptyText
-
-export const transformLocationsForTable = (locations) =>
-  locations ? locations.map((location) => location.name).join(', ') : ''
-
 export const transformBoolToRadioOption = (boolean) =>
   boolean ? OPTION_YES : OPTION_NO
-
-export const transformBoolToRadioOptionWithNullCheck = (boolean) =>
-  boolean === null ? null : transformBoolToRadioOption(boolean)
 
 export const transformRadioOptionToBool = (radioOption) =>
   radioOption === OPTION_YES
 
-const transformRadioOptionToInvertedBool = (radioOption) =>
-  radioOption === null ? null : radioOption === OPTION_NO
-
-export const transformBoolToInvertedRadioOption = (boolean) =>
-  boolean ? OPTION_NO : OPTION_YES
-
-export const transformBoolToInvertedRadioOptionWithNullCheck = (boolean) =>
-  boolean === null ? null : transformBoolToInvertedRadioOption(boolean)
-
 export const transformRadioOptionToBoolWithNullCheck = (boolean) =>
   boolean === null ? null : transformRadioOptionToBool(boolean)
-
-const setReferralSourceEvent = (values) => {
-  const {
-    referral_source_activity,
-    referral_source_activity_marketing,
-    referral_source_activity_website,
-    referral_source_activity_event,
-  } = values
-  return checkIfItemHasValue(referral_source_activity_marketing) ||
-    checkIfItemHasValue(referral_source_activity_website) ||
-    checkIfItemHasValue(referral_source_activity)
-    ? referral_source_activity_event
-    : ''
-}
-
-const setReferralSourceAdviser = (currentAdviser, values) => {
-  const { is_referral_source, referral_source_adviser } = values
-  return is_referral_source === 'yes'
-    ? currentAdviser
-    : checkIfItemHasValue(referral_source_adviser?.value)
-}
-
-const setConditionalArrayValue = (radioValue, array) =>
-  transformRadioOptionToBool(radioValue) ? array.map((x) => x.value) : []
-
-const setConditionalStringValue = (inputValue, investmentValue) =>
-  transformRadioOptionToBool(inputValue) ? investmentValue : null
-
-const setSiteDecidedSubValues = (
-  site_decided,
-  address1,
-  address2,
-  city,
-  postcode
-) => {
-  return transformRadioOptionToBool(site_decided)
-    ? {
-        address_1: address1,
-        address_2: address2,
-        address_town: city,
-        address_postcode: postcode,
-      }
-    : {
-        address_1: '',
-        address_2: '',
-        address_town: '',
-        address_postcode: '',
-      }
-}
-
-export const checkLandDate = (estimatedLandDate) => {
-  if (estimatedLandDate.day === '') {
-    estimatedLandDate.day = '01'
-  }
-  return transformDateObjectToDateString(estimatedLandDate)
-}
-
-export const setGVAMessage = (project) => {
-  const { foreignEquityInvestment, sector } = project
-  if (!foreignEquityInvestment && !sector) {
-    return 'Add capital expenditure value and primary sector (investment project summary) to calculate GVA'
-  }
-
-  if (!foreignEquityInvestment) {
-    return 'Add capital expenditure value and click "Save" to calculate GVA'
-  }
-
-  if (!sector) {
-    return 'Add primary sector (investment project summary) to calculate GVA'
-  }
-}
-
-export const transformProjectSummaryForApi = ({
-  projectId,
-  currentAdviser,
-  values,
-}) => {
-  const {
-    name,
-    description,
-    anonymous_description,
-    estimated_land_date,
-    investment_type,
-    fdi_type,
-    actual_land_date,
-    sector,
-    likelihood_to_land,
-    investor_type,
-    level_of_involvement,
-    specific_programme,
-    business_activities,
-    other_business_activity,
-    client_contacts,
-    referral_source_activity,
-    referral_source_activity_marketing,
-    referral_source_activity_website,
-  } = values
-
-  return {
-    id: projectId,
-    name,
-    description,
-    anonymous_description,
-    estimated_land_date: checkLandDate(estimated_land_date),
-    investment_type: checkIfItemHasValue(investment_type),
-    fdi_type: checkIfItemHasValue(fdi_type?.value),
-    actual_land_date: transformDateObjectToDateString(actual_land_date),
-    sector: checkIfItemHasValue(sector?.value),
-    likelihood_to_land: checkIfItemHasValue(likelihood_to_land?.value),
-    investor_type: checkIfItemHasValue(investor_type),
-    level_of_involvement: checkIfItemHasValue(level_of_involvement?.value),
-    specific_programme: checkIfItemHasValue(specific_programme?.value),
-    other_business_activity,
-    business_activities: business_activities.map((x) => x.value),
-    client_contacts: client_contacts.map((x) => x.value),
-    referral_source_activity: checkIfItemHasValue(referral_source_activity),
-    referral_source_activity_marketing: checkIfItemHasValue(
-      referral_source_activity_marketing
-    ),
-    referral_source_activity_website: checkIfItemHasValue(
-      referral_source_activity_website
-    ),
-    referral_source_activity_event: setReferralSourceEvent(values),
-    referral_source_adviser: setReferralSourceAdviser(currentAdviser, values),
-  }
-}
-
-export const transformProjectRequirementsForApi = ({ projectId, values }) => {
-  const {
-    actual_uk_regions,
-    address1,
-    address2,
-    city,
-    client_considering_other_countries,
-    client_requirements,
-    competitor_countries,
-    delivery_partners,
-    postcode,
-    site_decided,
-    strategic_drivers,
-    uk_region_locations,
-  } = values
-
-  const siteDecidedObject = setSiteDecidedSubValues(
-    site_decided,
-    address1,
-    address2,
-    city,
-    postcode
-  )
-
-  const requirementsValues = {
-    id: projectId,
-    actual_uk_regions: setConditionalArrayValue(
-      site_decided,
-      actual_uk_regions
-    ),
-    client_considering_other_countries: transformRadioOptionToBoolWithNullCheck(
-      client_considering_other_countries
-    ),
-    client_requirements,
-    competitor_countries: setConditionalArrayValue(
-      client_considering_other_countries,
-      competitor_countries
-    ),
-    delivery_partners: delivery_partners.map((x) => x.value),
-    site_decided: transformRadioOptionToBoolWithNullCheck(site_decided),
-    strategic_drivers: strategic_drivers.map((x) => x.value),
-    uk_region_locations: uk_region_locations.map((x) => x.value),
-  }
-
-  return { ...siteDecidedObject, ...requirementsValues }
-}
-
-export const transformProjectValueForApi = ({ projectId, values }) => {
-  const {
-    average_salary,
-    client_cannot_provide_foreign_investment,
-    client_cannot_provide_total_investment,
-    export_revenue,
-    fdi_value,
-    foreign_equity_investment,
-    government_assistance,
-    gross_value_added,
-    new_tech_to_uk,
-    non_fdi_r_and_d_budget,
-    number_new_jobs,
-    number_safeguarded_jobs,
-    r_and_d_budget,
-    total_investment,
-  } = values
-
-  return {
-    id: projectId,
-    average_salary: checkIfItemHasValue(average_salary),
-    client_cannot_provide_foreign_investment:
-      transformRadioOptionToInvertedBool(
-        client_cannot_provide_foreign_investment
-      ),
-    client_cannot_provide_total_investment: transformRadioOptionToInvertedBool(
-      client_cannot_provide_total_investment
-    ),
-    export_revenue: transformRadioOptionToBoolWithNullCheck(export_revenue),
-    fdi_value: checkIfItemHasValue(fdi_value),
-    foreign_equity_investment: setConditionalStringValue(
-      client_cannot_provide_foreign_investment,
-      foreign_equity_investment
-    ),
-    government_assistance: transformRadioOptionToBoolWithNullCheck(
-      government_assistance
-    ),
-    gross_value_added: setConditionalStringValue(
-      client_cannot_provide_foreign_investment,
-      gross_value_added
-    ),
-    new_tech_to_uk: transformRadioOptionToBoolWithNullCheck(new_tech_to_uk),
-    non_fdi_r_and_d_budget: transformRadioOptionToBoolWithNullCheck(
-      non_fdi_r_and_d_budget
-    ),
-    number_new_jobs: checkIfItemHasValue(number_new_jobs),
-    number_safeguarded_jobs: checkIfItemHasValue(number_safeguarded_jobs),
-    r_and_d_budget: transformRadioOptionToBoolWithNullCheck(r_and_d_budget),
-    total_investment: setConditionalStringValue(
-      client_cannot_provide_total_investment,
-      total_investment
-    ),
-  }
-}
 
 export const transformProjectStatusForApi = ({ project, values }) => ({
   id: project.id,
@@ -378,14 +122,6 @@ export const transformProjectStageForApi = (values) => {
 
 export const transformFdiType = (FDI, FDIType) =>
   FDIType ? FDI + ', ' + FDIType : FDI + ', None'
-
-export const transformBusinessActivity = (
-  businessActivity,
-  otherBusinessActivity
-) =>
-  otherBusinessActivity
-    ? transformArray(businessActivity) + ', ' + otherBusinessActivity
-    : transformArray(businessActivity)
 
 export const transformFdiRAndDProject = (project) => {
   if (project.associatedNonFdiRAndDProject) {
@@ -499,157 +235,3 @@ export const transformPropositionToListItem = ({
       </>
     ),
 })
-
-const transformNonFdiProjectToListItem = (projectId) => (project) => {
-  const {
-    id,
-    name,
-    project_code,
-    stage,
-    investment_type,
-    status,
-    investor_company,
-    estimated_land_date,
-    sector,
-  } = project
-
-  const badges = [
-    { text: stage.name },
-    { text: investment_type.name },
-    { text: status },
-  ]
-
-  const metadata = [
-    { label: 'Investor', value: investor_company.name },
-    { label: 'Sector', value: sector ? sector.name : '' },
-    {
-      label: 'Estimated land date',
-      value: estimated_land_date && format(estimated_land_date, 'MMMM yyyy'),
-    },
-  ].filter((metadata) => metadata.value)
-
-  return {
-    id,
-    headingUrl: urls.investments.projects.editAssociatedProject(projectId, id),
-    headingText: name,
-    subheading: `Project code ${project_code}`,
-    badges,
-    metadata,
-  }
-}
-
-export const transformNonFdiResponseToCollection = (
-  { count, results = [] },
-  projectId
-) => ({
-  count,
-  results: results.map(transformNonFdiProjectToListItem(projectId)),
-})
-
-export const checkIfAssociatedProjectExists = (hasAssociatedProject) =>
-  hasAssociatedProject ? 'Update associated project' : 'Add associated project'
-
-const transformCompanyToListItem =
-  (projectId) =>
-  ({
-    id,
-    name,
-    sector,
-    uk_region,
-    trading_names,
-    address,
-    modified_on,
-    headquarter_type,
-    global_headquarters,
-    latest_interaction_date,
-  } = {}) => {
-    const metadata = [
-      {
-        label: 'Sector',
-        value: get(sector, 'name'),
-      },
-      {
-        label: 'Global HQ',
-        value: get(global_headquarters, 'name'),
-      },
-      {
-        label: labels.address.companyAddress,
-        value: addressToString(address),
-      },
-    ]
-
-    if (trading_names && trading_names.length) {
-      metadata.push({
-        label: labels.hqLabels.trading_names,
-        value: trading_names.join(', '),
-      })
-    }
-
-    if (latest_interaction_date) {
-      metadata.push({
-        label: 'Last interaction date',
-        value: format(latest_interaction_date),
-      })
-    }
-
-    if (headquarter_type) {
-      metadata.push({
-        label: 'Headquarter type',
-        value: labels.hqLabels[get(headquarter_type, 'name')],
-      })
-    }
-
-    const badges = [
-      { text: get(address, 'country.name') },
-      { text: get(uk_region, 'name') },
-      { text: labels.hqLabels[get(headquarter_type, 'name')] },
-    ].filter((item) => item.text)
-
-    return {
-      id,
-      subheading: modified_on
-        ? `Updated on ${formatMediumDateTime(modified_on)}`
-        : undefined,
-      headingText: name,
-      headingUrl: urls.investments.projects.editRecipientCompany(projectId, id),
-      badges,
-      metadata: metadata.filter((item) => item.value),
-    }
-  }
-
-export const transformResponseToCompanyCollection = (
-  projectId,
-  { count, results = [] }
-) => ({
-  count,
-  results: results.map(transformCompanyToListItem(projectId)),
-})
-
-export const checkIfRecipientCompanyExists = (hasRecipientCompany) =>
-  hasRecipientCompany ? 'Update recipient company' : 'Add recipient company'
-
-export const transformValuesToArray = (values) => {
-  let result = []
-
-  Object.entries(values).forEach(([key, value]) => {
-    const prefix = key.split('_')[0]
-    const index = key.split('_')[1]
-    result[index] = {
-      ...result[index],
-      [prefix]: value.value || value,
-    }
-  })
-  return result.filter((item) => item !== null)
-}
-
-export const transformTeamMembersForFieldAddAnother = (teamMembers = []) => {
-  const teamMemberArray = teamMembers.map((teamMember, index) => ({
-    ['adviser' + `_${index}`]: {
-      value: teamMember.adviser?.id,
-      label: teamMember.adviser?.name,
-    },
-    ['role' + `_${index}`]: teamMember.role,
-  }))
-
-  return transformArrayToObject(teamMemberArray)
-}
