+++ src/client/routes.js
@@ -24,25 +24,25 @@ import EditProjectStatus from './modules/Investments/Projects/EditProjectStatus'
 import ObjectiveAdd from './modules/Companies/AccountManagement/Objective/ObjectiveAdd'
 import ObjectiveEdit from './modules/Companies/AccountManagement/Objective/ObjectiveEdit'
 import ArchivedObjectives from './modules/Companies/AccountManagement/ArchivedObjectives'
-import FindAssociatedProject from './modules/Investments/Projects/FindAssociatedProject'
-import EditAssociatedProject from './modules/Investments/Projects/EditAssociatedProject'
-import FindRecipientCompany from './modules/Investments/Projects/FindRecipientCompany'
-import EditRecipientCompany from './modules/Investments/Projects/EditRecipientCompany'
+import FindAssociatedProject from './modules/Investments/Projects/Details/EditAssociatedProject/FindAssociatedProject'
+import EditAssociatedProject from './modules/Investments/Projects/Details/EditAssociatedProject/EditAssociatedProject'
+import FindRecipientCompany from './modules/Investments/Projects/Details/EditRecipientCompany/FindRecipientCompany'
+import EditRecipientCompany from './modules/Investments/Projects/Details/EditRecipientCompany/EditRecipientCompany'
 import EditOneList from './modules/Companies/CoreTeam/EditOneList'
-import ProjectDetails from './modules/Investments/Projects/ProjectDetails'
-import ProjectTeam from './modules/Investments/Projects/ProjectTeam'
-import EditProjectSummary from './modules/Investments/Projects/EditProjectSummary'
-import EditProjectRequirements from './modules/Investments/Projects/EditProjectRequirements'
-import EditProjectValue from './modules/Investments/Projects/EditProjectValue'
+import ProjectDetails from './modules/Investments/Projects/Details/ProjectDetails'
+import ProjectTeam from './modules/Investments/Projects/Team/ProjectTeam'
+import EditProjectSummary from './modules/Investments/Projects/Details/EditProjectSummary'
+import EditProjectRequirements from './modules/Investments/Projects/Details/EditProjectRequirements'
+import EditProjectValue from './modules/Investments/Projects/Details/EditProjectValue'
 import ProjectEvaluation from './modules/Investments/Projects/ProjectEvaluation'
 import ProjectInteractions from './modules/Investments/Projects/ProjectInteractions'
 import ProjectPropositions from './modules/Investments/Projects/ProjectPropositions'
-import EditClientRelationshipManagement from './modules/Investments/Projects/EditClientRelationshipManagement'
-import EditProjectManagement from './modules/Investments/Projects/EditProjectManagement'
-import EditTeamMembers from './modules/Investments/Projects/EditTeamMembers'
+import EditClientRelationshipManagement from './modules/Investments/Projects/Team/EditClientRelationshipManagement'
+import EditProjectManagement from './modules/Investments/Projects/Team/EditProjectManagement'
+import EditTeamMembers from './modules/Investments/Projects/Team/EditTeamMembers'
 import ProjectEditHistory from './modules/Investments/Projects/ProjectEditHistory'
-import ProjectEvidence from './modules/Investments/Projects/ProjectEvidence'
-import DeleteProjectDocument from './modules/Investments/Projects/DeleteProjectDocument'
+import ProjectEvidence from './modules/Investments/Projects/Evidence/ProjectEvidence'
+import DeleteProjectDocument from './modules/Investments/Projects/Evidence/DeleteProjectDocument'
 
 const routes = {
   companies: [
